requirements.txt
black==23.3.0
click==8.1.3
gitdb==4.0.10
GitPython==3.1.31
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pluggy==1.0.0
pytest==7.3.1
smmap==5.0.0
